+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # OMC3 Changelog
 
+#### 2023-04-20 - v0.8.0 - _jdilly_
+
+
+
+- Fix:
+  - Changed all `pandas`/`tfs-pandas` `append()` and `join()` to `concat()`
+  - Moved `InputFiles` into separate module
+
 #### 2023-03-16 - v0.7.2 - _jdilly_
 
 - Fix: